feat: centralize budget creation
