+++ q-srfm/src/pages/DashboardPage.vue
@@ -356,14 +356,14 @@ import IncomeVsExpensesCard from '../components/charts/IncomeVsExpensesCard.vue'
 import type { Transaction, Budget, IncomeTarget, BudgetCategoryTrx, BudgetCategory } from '../types';
 import { EntityType } from '../types';
 import version from '../version';
-import { toDollars, toCents, formatCurrency, adjustTransactionDate, todayISO, currentMonthISO } from '../utils/helpers';
+import { toDollars, toCents, formatCurrency, todayISO, currentMonthISO } from '../utils/helpers';
 import { useAuthStore } from '../store/auth';
 import { useBudgetStore } from '../store/budget';
 import { useMerchantStore } from '../store/merchants';
 import { useFamilyStore } from '../store/family';
 import debounce from 'lodash/debounce';
 import { v4 as uuidv4 } from 'uuid';
-import { DEFAULT_BUDGET_TEMPLATES } from '../constants/budgetTemplates';
+import { createBudgetForMonth } from '../utils/budget';
 
 // Structured logger for this page
 const DBG = '[Dashboard]';
@@ -851,155 +851,6 @@ async function loadBudgets() {
   }
 }
 
-async function createBudgetForMonth(month: string, familyId: string, ownerUid: string, entityId: string): Promise<Budget> {
-  const budgetId = `${ownerUid}_${entityId}_${month}`;
-  const existingBudget = await dataAccess.getBudget(budgetId);
-  if (existingBudget) {
-    return existingBudget;
-  }
-
-  const entity = familyStore.family?.entities?.find((e) => e.id === entityId);
-  const templateBudget = entity?.templateBudget;
-
-  if (templateBudget && templateBudget.categories.length > 0) {
-    const newBudget: Budget = {
-      familyId: familyId,
-      entityId: entityId,
-      month: month,
-      incomeTarget: 0,
-      categories: templateBudget.categories.map((cat) => ({
-        ...cat,
-        carryover: cat.isFund ? 0 : 0,
-      })),
-      transactions: [],
-      label: `Template Budget for ${month}`,
-      merchants: [],
-      budgetId: budgetId,
-    };
-
-    await dataAccess.saveBudget(budgetId, newBudget);
-    budgetStore.updateBudget(budgetId, newBudget);
-    budgets.value.push(newBudget);
-    return newBudget;
-  }
-
-  if (entity && DEFAULT_BUDGET_TEMPLATES[entity.type]) {
-    const predefinedTemplate = DEFAULT_BUDGET_TEMPLATES[entity.type];
-    const newBudget: Budget = {
-      familyId: familyId,
-      entityId: entityId,
-      month: month,
-      incomeTarget: 0,
-      categories: predefinedTemplate.categories.map((cat) => ({
-        ...cat,
-        carryover: cat.isFund ? 0 : 0,
-      })),
-      transactions: [],
-      label: `Default ${entity.type} Budget for ${month}`,
-      merchants: [],
-      budgetId: budgetId,
-    };
-
-    await dataAccess.saveBudget(budgetId, newBudget);
-    budgetStore.updateBudget(budgetId, newBudget);
-    budgets.value.push(newBudget);
-    return newBudget;
-  }
-
-  const availableBudgets = Array.from(budgetStore.budgets.values()).sort((a, b) => a.month.localeCompare(b.month));
-  let sourceBudget;
-
-  const previousBudgets = availableBudgets.filter((b) => b.month < month && b.entityId === entityId);
-  if (previousBudgets.length > 0) {
-    sourceBudget = previousBudgets[previousBudgets.length - 1];
-  } else {
-    const futureBudgets = availableBudgets.filter((b) => b.month > month && b.entityId === entityId);
-    if (futureBudgets.length > 0) {
-      sourceBudget = futureBudgets[0];
-    }
-  }
-
-  if (sourceBudget) {
-    const [newYear, newMonthNum] = month.split('-').map(Number);
-    const [sourceYear, sourceMonthNum] = sourceBudget.month.split('-').map(Number);
-    const isFutureMonth = newYear > sourceYear || (newYear === sourceYear && newMonthNum > sourceMonthNum);
-
-    let newCarryover: Record<string, number> = {};
-    if (isFutureMonth) {
-      newCarryover = dataAccess.calculateCarryOver(sourceBudget);
-    }
-
-    const newBudget: Budget = {
-      familyId: familyId,
-      entityId: entityId,
-      month: month,
-      incomeTarget: sourceBudget.incomeTarget,
-      categories: sourceBudget.categories.map((cat) => ({
-        ...cat,
-        carryover: cat.isFund ? newCarryover[cat.name] || 0 : 0,
-      })),
-      label: sourceBudget.label || `Budget for ${month}`,
-      merchants: sourceBudget.merchants || [],
-      transactions: [],
-      budgetId: budgetId,
-    };
-
-    const recurringTransactions: Transaction[] = [];
-    if (sourceBudget.transactions) {
-      const recurringGroups: Record<string, Transaction[]> = sourceBudget.transactions.reduce(
-        (groups, trx) => {
-          if (!trx.deleted && trx.recurring) {
-            const key = `${trx.merchant}-${trx.amount}-${trx.recurringInterval}-${trx.userId}-${trx.isIncome}`;
-            if (!groups[key]) {
-              groups[key] = [];
-            }
-            groups[key].push(trx);
-          }
-          return groups;
-        },
-        {} as Record<string, Transaction[]>,
-      );
-
-      Object.values(recurringGroups).forEach((group) => {
-        const firstInstance = group.sort((a, b) => new Date(a.date).getTime() - new Date(b.date).getTime())[0];
-        if (firstInstance.recurringInterval === 'Monthly') {
-          const newDate = adjustTransactionDate(firstInstance.date, month, 'Monthly');
-          recurringTransactions.push({
-            ...firstInstance,
-            id: uuidv4(),
-            date: newDate,
-            budgetMonth: month,
-            entityId: entityId,
-          });
-        }
-      });
-    }
-
-    newBudget.transactions = recurringTransactions;
-    await dataAccess.saveBudget(budgetId, newBudget);
-    budgetStore.updateBudget(budgetId, newBudget);
-    budgets.value.push(newBudget);
-    return newBudget;
-  }
-
-  const defaultBudget: Budget = {
-    familyId: familyId,
-    entityId: entityId,
-    month: month,
-    incomeTarget: 0,
-    categories: [
-      { name: 'Income', target: 0, isFund: false, group: 'Income' },
-      { name: 'Miscellaneous', target: 0, isFund: false, group: 'Expenses' },
-    ],
-    transactions: [],
-    label: `Default Budget for ${month}`,
-    merchants: [],
-    budgetId: budgetId,
-  };
-  await dataAccess.saveBudget(budgetId, defaultBudget);
-  budgetStore.updateBudget(budgetId, defaultBudget);
-  return defaultBudget;
-}
 
 async function createDefaultBudget() {
   const user = auth.user;
